refactor predictions and clean up the logs
