+++ plantseg/predictions/functional/array_dataset.py
