+++ plantseg/predictions/functional/predictions.py
@@ -4,7 +4,7 @@ from numpy.typing import ArrayLike
 from pytorch3dunet.unet3d import utils
 
 from plantseg.dataprocessing.functional.dataprocessing import fix_input_shape
-from plantseg.predictions.utils import get_dataset_config, get_model_config, get_predictor_config, set_device
+from plantseg.predictions.functional.utils import get_dataset_config, get_model_config, get_predictor_config, set_device
 
 
 def unet_predictions(raw: ArrayLike,
