+++ plantseg/predictions/predict.py
@@ -3,7 +3,7 @@ from pytorch3dunet.unet3d import utils
 from plantseg.io.io import load_shape
 from plantseg.pipeline import gui_logger
 from plantseg.pipeline.steps import GenericPipelineStep
-from plantseg.predictions.utils import get_dataset_config, get_model_config, get_predictor_config, set_device
+from plantseg.predictions.functional.utils import get_dataset_config, get_model_config, get_predictor_config, set_device
 
 
 def _check_patch_size(paths, patch_size):
@@ -62,7 +62,7 @@ class UnetPredictions(GenericPipelineStep):
         model, model_config, model_path = get_model_config(model_name, model_update=model_update, version=version)
         utils.load_checkpoint(model_path, model)
 
-        device = set_device(device)
+        device = set_device(device, device_id=0)
         model = model.to(device)
 
         predictor, predictor_config = get_predictor_config(model_name)
